refactoring

+++ Core/Inc/FlightController/PID.h
@@ -2,6 +2,7 @@
 // Created by Dmytro Hrachov on 01.05.2026.
 //
 #pragma once
+#include <cstdint>
 
 struct  PID
 {
@@ -18,6 +19,6 @@ struct  PID
 class PID_Controller
 {
 public:
-    static float Update(PID *pid, float target, float measured, float dt);
-    static float UpdateAngleWithGyroD(PID *pid, float targetAngleDeg, float measuredAngleDeg, float gyroDegPerSec, float dt);
+    static int16_t Update(PID *pid, float target, float measured, float dt);
+    static int16_t UpdateAngleWithGyroD(PID *pid, float targetAngleDeg, float measuredAngleDeg, float gyroDegPerSec, float dt);
 };
